+++ mod.ts
@@ -1,2 +1,2 @@
-export { importw } from "./src/importw.ts";
+export { importw, workerSymbol } from "./src/importw.ts";
 export { importw as default } from "./src/importw.ts";